Added sprite sheets.
Do note that when i tested the program with my test sprite sheet i did
not get the desired result of a venesaur picture. Instead i was given a
section of the picture. I do not know if this is a problem with the code
of a problem with the test picture i chose if it is exactly 64x64 size
sprites or if there are any offsets